# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Postmark Service API",
    "description": "Email sending and tracking service built on Postmark. Handles email delivery via the broadcast message stream, webhook processing for delivery events, and integrates with a runs-service for cost tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://postmark.mcpfactory.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:3010",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "required": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "OrgEmailsResponse": {
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "emails": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "messageId": {
                  "type": "string",
                  "nullable": true
                },
                "to": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "submittedAt": {
                  "type": "string",
                  "nullable": true
                },
                "runId": {
                  "type": "string",
                  "nullable": true
                },
                "errorCode": {
                  "type": "number",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "messageId",
                "to",
                "subject",
                "submittedAt",
                "runId",
                "errorCode"
              ]
            }
          }
        },
        "required": [
          "orgId",
          "count",
          "emails"
        ]
      },
      "RunEmailsResponse": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "emails": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "messageId": {
                  "type": "string",
                  "nullable": true
                },
                "to": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "submittedAt": {
                  "type": "string",
                  "nullable": true
                },
                "success": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "messageId",
                "to",
                "subject",
                "submittedAt",
                "success"
              ]
            }
          }
        },
        "required": [
          "runId",
          "total",
          "emails"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/status/{messageId}": {
      "get": {
        "summary": "Get email status",
        "description": "Get the full delivery status of an email by its Postmark message ID",
        "tags": [
          "Email Status"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Postmark message ID"
            },
            "required": true,
            "description": "Postmark message ID",
            "name": "messageId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Email status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStatus"
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/status/by-org/{orgId}": {
      "get": {
        "summary": "Get emails by organization",
        "description": "Get recent emails for an organization",
        "tags": [
          "Email Status"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Organization ID"
            },
            "required": true,
            "description": "Organization ID",
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Max results (default: 50)"
            },
            "required": false,
            "description": "Max results (default: 50)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Organization emails",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgEmailsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/status/by-run/{runId}": {
      "get": {
        "summary": "Get emails by run",
        "description": "Get all emails for a specific run",
        "tags": [
          "Email Status"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Run ID"
            },
            "required": true,
            "description": "Run ID",
            "name": "runId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Run emails",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEmailsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "post": {
        "summary": "Batch status lookup by lead and email",
        "description": "Check delivery status for lead+email pairs. Returns campaign-scoped (optional), brand-scoped, and global results.",
        "tags": [
          "Email Status"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-item status results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get aggregated stats",
        "description": "Get aggregated email stats optionally filtered by runIds, orgId, brandId, campaignId, workflowSlug, featureSlug, workflowDynastySlug, and/or featureDynastySlug. Dynasty slug filters resolve to all versioned slugs via the respective service. When no filters are provided, returns stats across all sendings. When groupBy is provided, returns grouped results. Requires x-org-id and x-user-id headers.",
        "tags": [
          "Email Status"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated run IDs"
            },
            "required": false,
            "description": "Comma-separated run IDs",
            "name": "runIds",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by organization ID"
            },
            "required": false,
            "description": "Filter by organization ID",
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by brand ID"
            },
            "required": false,
            "description": "Filter by brand ID",
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by campaign ID"
            },
            "required": false,
            "description": "Filter by campaign ID",
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow slug"
            },
            "required": false,
            "description": "Filter by workflow slug",
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature slug"
            },
            "required": false,
            "description": "Filter by feature slug",
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by workflow dynasty slug (resolves to all versioned slugs via workflow-service)"
            },
            "required": false,
            "description": "Filter by workflow dynasty slug (resolves to all versioned slugs via workflow-service)",
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by feature dynasty slug (resolves to all versioned slugs via features-service)"
            },
            "required": false,
            "description": "Filter by feature dynasty slug (resolves to all versioned slugs via features-service)",
            "name": "featureDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "brandId",
                "campaignId",
                "workflowSlug",
                "featureSlug",
                "workflowDynastySlug",
                "featureDynastySlug",
                "leadEmail"
              ],
              "description": "Group results by dimension"
            },
            "required": false,
            "description": "Group results by dimension",
            "name": "groupBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated stats (flat or grouped depending on groupBy parameter)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/StatsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GroupedStatsResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}